refactor upload function
